Reduced iam privileges (#4)
* Reduced log privileges

Used role had too many permissions. With this change only the least required
privileger are granted.

S3 access is not required in the exporter role but on the S3 bucket policy for
logs principal. This is terraform snippet that does the trick

{
    "Action": "s3:GetBucketAcl",
    "Effect": "Allow",
    "Resource": "${aws_s3_bucket.log.arn}",
    "Principal": { "Service": "logs.${data.aws_region.current.name}.amazonaws.com" }
},
{
    "Action": "s3:PutObject" ,
    "Effect": "Allow",
    "Resource": "${aws_s3_bucket.log.arn}/*",
    "Condition": { "StringEquals": { "s3:x-amz-acl": "bucket-owner-full-control" } },
    "Principal": { "Service": "logs.${data.aws_region.current.name}.amazonaws.com" }

* Added further wait time via retry loop

Retry loop it's limited on the number of iterations to avoid a long running
function.

Also ssm_paramter update has been moved into the try in order to avoid storing
a timestamp which hasn't happened due to reaching max_retries.